Created DataRetrievalInterceptor to replace AuditLogFilter for Rest-Api
